Some controller cleanup
